Bugfix/function detection (#14)
* Fixes inconsistency with local/global inline functions
A change was made that fixes the problem that a
local/global inline function appeared as a variable and a function
at the same time in the outline view (additionally: badly structured).
Now such a function definition is show just as a function.

Additionally the regex for constants was simplified a bit.

* Anonymous function don't break the structure anymore
If you passed an anonymous function as a parameter into another
function, the structure was broken. This is fixed now.
Example: Sort(array, Function(a, b) Return(a < b) EndFunction)

* Fixed broken structure caused by comments
When multiple variables are declared on one line
it was possible that a comment at the end of this
line caused the structure to be broken like in this example:
`Local t9 ;, t10`
This was fixed now.

* Fixed broken structure caused by comments
Ignoring /* comments from the last commit did not work.

* Improved some regular expressions
The regular expression for detecting functions were improved.