Merge dev to master (#18)
* Fixed linking error in README.md

* Added more snippets

* 8.0.3

* feat: Improved textmate grammar (#10)

- Added Syntax Highlighting for MulitLine strings (see https://www.hollywood-mal.com/docs/html/hollywood/PrgTypeString.html)

- Added more escaped characters for better Syntax Highlighting

- Improved some grammer identifiers

* Feature/function detection parser (#11)

* Massively improved syntax highlighting detection
for standard Hollywood function definition (Function - EndFunction)

Changed some keyword assignments for Syntax Highlighting

* First experimental version of a LanguageProvider.

In this case it is a DocumentSymbolProvider in order to get the
outline view and other helpers for the defined variables
and functions.
There is still a lot to fix and improve ...

* Added compilation output folder to .gitignore

* Implemented proper calculation of the position
range for function and variable declaration which
also fixed a bug of wrong structures in the outline view

* Added support for constants; code cleanup

Recogonition of self defined constants (Const #MYCONSTANT) is supported.

The code was cleaned up a bit.

* Small optimizations

Just some small code optimizations.

* Multiline comment detection added
Now multiline comments are detected as well, so you
don't get function and variable definitions in the outline view etc.
in commented sections anymore.

* Fixes inconsistency with local/global inline functions (#13)

A change was made that fixes the problem that a
local/global inline function appeared as a variable and a function
at the same time in the outline view (additionally: badly structured).
Now such a function definition is show just as a function.

Additionally the regex for constants was simplified a bit.

* Bugfix/function detection (#14)

* Fixes inconsistency with local/global inline functions
A change was made that fixes the problem that a
local/global inline function appeared as a variable and a function
at the same time in the outline view (additionally: badly structured).
Now such a function definition is show just as a function.

Additionally the regex for constants was simplified a bit.

* Anonymous function don't break the structure anymore
If you passed an anonymous function as a parameter into another
function, the structure was broken. This is fixed now.
Example: Sort(array, Function(a, b) Return(a < b) EndFunction)

* Fixed broken structure caused by comments
When multiple variables are declared on one line
it was possible that a comment at the end of this
line caused the structure to be broken like in this example:
`Local t9 ;, t10`
This was fixed now.

* Fixed broken structure caused by comments
Ignoring /* comments from the last commit did not work.

* Improved some regular expressions
The regular expression for detecting functions were improved.

* Feature/go to definition (#16)

* Begin of "Go to Definition" implementation
This is the inital code/basic structure of the "Go to definition"
functionality implemented with a DefinitionProvider

* "Go to definition" for normal functions
Now it is possible to use the "Go to definition" functionality for
normal Hollywood function (multiline and oneliner),
if the definition is in the same document.

* "Go to definition" extended for variables
The "Go to definiton" functionality now also works for variables
defined as Local or Global.

* Go to definition improved and extented
It also works for constants and inline functions now.
Additionally there were many fixes made that where reported by mrupp.
This commit also fixes #6 through adding a word pattern in the
language-configuration.json. This was needed so the selected word
when pressing F12 is taking special characters like #, ! and $
into account.

* Commented lines are ignored now
A definition that is commented out is ignored now.
A helper library (utils.ts) was created to extract the function
for finding commented lines, so this functionality can be reused.
Therefore the documentSymbolProvider was modified, too.

* Support for one line variable definitions added
Now the "Go to definiton" feature does work for one line definitions:
Local t1, t2, t3 = 1, 2, 3

Additionally the code for definitionProvider and documentSymbolProvider
was made more robust regarding single line and multi line comments
within a variable definition list, so those examples won't break
the structure and functionality:
Local t1 /* t2 */, t3
Local t4 , ;t5

* just a small code cleanup

* Feature/8.0.4rc (#17)

* Added and updated documentation for 8.0.4 release

* Improved syntax highlighting in tables
The syntax highlighting for values in tables was improved a bit (#12).

* Different icon for global variables
Global variables have a different icon to better distinquish from
local variables now.

* Improved README.md
Notes and images regarding the supported symbols were added.

Some typos has been fixed.